Renamed duplicate HomepageStatistics component to HomepageStatisticsLegacy and moved it to component folder to avoid confusion
