+++ src/components/Homepage/HomepageStatisticsLegacy.vue
